+++ massively/massively/src/core/eval.rs
@@ -189,7 +189,7 @@ impl<ValuesExpr, IndicesExpr, Item> DeviceExpr<Item> for PermuteExpr<ValuesExpr,
 where
     Item: CubeType + 'static,
     ValuesExpr: DeviceExpr<Item>,
-    IndicesExpr: DeviceExpr<u32>,
+    IndicesExpr: DeviceExpr<usize>,
 {
 }
 
@@ -347,7 +347,7 @@ macro_rules! define_eval {
             Item: CubeType + 'static,
             $( $leaf: CubePrimitive, )+
             ValuesExpr: $trait_name<Item, $( $leaf ),+>,
-            IndicesExpr: $trait_name<u32, $( $leaf ),+>,
+            IndicesExpr: $trait_name<usize, $( $leaf ),+>,
         {
             fn $method(
                 $( $slot: &[$leaf], )+
@@ -355,7 +355,7 @@ macro_rules! define_eval {
                 index: usize,
             ) -> Item {
                 let gathered = IndicesExpr::$method($( $slot, )+ slot_offsets, index);
-                ValuesExpr::$method($( $slot, )+ slot_offsets, gathered as usize)
+                ValuesExpr::$method($( $slot, )+ slot_offsets, gathered)
             }
         }
